refactor scraper from Module to Class
